# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API documentiation for pargar team",
    "version": "1.0.0",
    "title": "Pargar Swagger"
  },
  "host": "pargar.herokuapp.com",
...
  "schemes": [
    "http",
    "ws"
  ],
  "paths": {
...
    "/chat/direct/contact/": {
      "get": {
        "tags": [
          "Direct"
        ],
        "summary": "List of Connecting User(has chat with them)",
        "operationId": "DirectMessage",
        "security": [
          {
            "Token": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "username": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "user_type": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden "
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}